requirements.txt
Flask==2.3.2
Werkzeug==2.3.8
python-dotenv
psycopg2-binary